PDJB-764: Refactor integration test page objects
