+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/updateGasSafetyJourneyPages/UploadGasCertFormPageUpdateGasSafety.kt
@@ -2,26 +2,15 @@ package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.updateGasS
 
 import com.microsoft.playwright.Page
 import uk.gov.communities.prsdb.webapp.controllers.UpdateGasSafetyController
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.Heading
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.PostForm
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.UploadCertFormBasePage
 import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.UploadGasCertStep
-import java.nio.file.Path
 
 class UploadGasCertFormPageUpdateGasSafety(
     page: Page,
     urlArguments: Map<String, String>,
-) : BasePage(
+) : UploadCertFormBasePage(
         page,
         UpdateGasSafetyController.UPDATE_GAS_SAFETY_ROUTE
             .replace("{propertyOwnershipId}", urlArguments["propertyOwnershipId"]!!) +
             "/${UploadGasCertStep.ROUTE_SEGMENT}",
-    ) {
-    val heading = Heading(page.locator("h1"))
-    val form = PostForm(page)
-
-    fun uploadGasCertificate(filePath: Path) {
-        page.setInputFiles("input[type=\"file\"]", filePath)
-        form.submit()
-    }
-}
+    )
